+++ src/ionic-tree-provider.ts
@@ -5,6 +5,7 @@ import * as fs from 'fs';
 import * as path from 'path';
 import { reviewProject, starterProject } from './recommendations';
 import { Recommendation } from './recommendation';
+import { Context } from './context-variables';
 
 export const ionicState = { view: undefined, skipAuth: false };
 
@@ -25,9 +26,10 @@ export class IonicTreeProvider implements vscode.TreeDataProvider<Recommendation
 
 	getChildren(element?: Recommendation): Thenable<Recommendation[]> {
 		if (!this.workspaceRoot) {
-			vscode.window.showInformationMessage('No dependency in empty workspace');
+			vscode.commands.executeCommand('setContext', Context.noProjectFound, true);
 			return Promise.resolve([]);
 		}
+		vscode.commands.executeCommand('setContext', Context.noProjectFound, false);
 
 		if (element) {
 			return Promise.resolve(element.children);